pyproject.toml
[project]
name = "autoneura"
version = "0.1.0"
description = ""
authors = [
    {name = "Erwin",email = "erwinwingyonata@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13.0"
dependencies = [
    "supervision (>=0.25.1,<0.26.0)",
    "ultralytics (>=8.3.99,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-box (>=7.3.2,<8.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "numpy (==2.0.2)",
    "protobuf (==5.29.4)",
    "decorator (>=5.2.1,<6.0.0)",
    "opt-einsum (>=3.4.0,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "python-doctr[torch,viz] (>=0.11.0,<0.12.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "rich (>=14.0.0,<15.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "paddlepaddle-cpu"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
priority = "supplemental"


[tool.poetry.dependencies]
